Move signing public, rename to keys
